MAINT: move ndarray methods to free functions
